rename commands to short-.*
