--- src/bin/club-find.ts
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-import { main } from './club-search';
-
-main();
