+++ src/bin/short-create.ts
