+++ src/bin/short-find.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { main } from './short-search';
+
+main();
