+++ src/bin/short-install.ts
