+++ src/bin/short-projects.ts
