+++ src/bin/short-search.ts
