+++ src/bin/short.ts
